requirements.txt
Flask~=1.1.2
pip~=20.0.2
wheel~=0.34.2
pytz~=2019.3
Jinja2~=2.11.2
Werkzeug~=0.16.1
itsdangerous~=1.1.0
click~=7.1.1
MarkupSafe~=1.1.1
aniso8601~=8.0.0
attrs~=19.3.0
jsonschema~=3.2.0
pyrsistent~=0.16.0
six~=1.14.0
setuptools~=46.1.3
gunicorn~=20.0.4
apis~=0.0.1
SQLAlchemy~=1.3.16
PyJWT~=1.7.1
flask-restplus==0.12.1
flask_sqlalchemy